Move linux inhibitor tests too
